StartNumbering renamed to NumberingForm
